prepare: split shared trait into another file
